addressed review comments and restructured
