+++ examples/image_text_to_text/models/gemma_vision/gemma4/gemma4_diss_cb_with_kv_share.py
@@ -208,11 +208,13 @@ def run(
         _resolve_lang_qpc_path(prefill_qpc_path, ("lang_prefill_qpc_path", "lang_qpc_path")),
         kv_dma_share=True,
         full_batch_size=full_batch_size,
+        cluster_id="prefill",
     )
     decode_session = QAICInferenceSession(
         _resolve_lang_qpc_path(decode_qpc_path, ("lang_decode_qpc_path", "lang_qpc_path")),
         kv_dma_share=True,
         full_batch_size=full_batch_size,
+        cluster_id="decode",
     )
 
     assert "image_idx" in prefill_session.binding_index_map, "image_idx not a compiled prefill input binding"
